fix: update validation schema to ensure positive integers for pagination

+++ src/lib/templates/validation.ts
@@ -55,8 +55,8 @@ export const listTemplatesQuerySchema = z.object({
   type: templateTypeEnum.optional(),
   language: languageEnum.optional(),
   channel: channelEnum.optional(),
-  page: z.coerce.number().int().positive().default(1),
-  limit: z.coerce.number().int().positive().max(100).default(50),
+  page: z.coerce.number().positive().default(1),
+  limit: z.coerce.number().positive().max(100).default(50),
 });
 
 export type ListTemplatesQuery = z.infer<typeof listTemplatesQuerySchema>;
